SpiritUtils library added for reusing useful functions. Currently has mode sequence functions for easy mode building.
